scripts/temp/requirements1.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
diskcache==5.6.3
dnspython==2.7.0
docopt==0.6.2
docx==0.2.4
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
faster-whisper==1.2.0
filelock==3.13.1
firebase_admin==7.0.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.3
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.2
lightning-utilities==0.14.3
llama_cpp_python==0.3.14
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
neo4j==5.28.1
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.22.1
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
primePy==1.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.5.12
pymongo==4.13.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyreadline3==3.5.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1+cpu
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1+cpu
torchmetrics==1.7.4
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
webrtcvad==2.0.10
websockets==15.0.1
yarl==1.20.1
